\name{big.eight}
\alias{big.eight}
\docType{data}
\title{Example dataset for Ross Ordering Pairing}
\description{
Dataset contains 8 stimuli listing the names of universities that made up the Big Nine conference (without Iowa).
}
\usage{data(big.nine)}
\format{
  A data frame with 8 observations on the following variable.
  \describe{
    \item{\code{University}}{is a data.frame with the following stimuli: 
	
	University of Chicago
	
	University of Illinois
	
	University of Michigan
	
	University of Minnesota
	
	Northwestern University
	
	Purdue University
	
	University of Wisconsin
	
	Indiana University }
  }
}
\examples{
data(big.eight)
ross.ordering(big.eight)}

