\name{big.nine}
\alias{big.nine}
\docType{data}
\title{Example dataset for Ross Ordering Pairing}
\description{
Dataset contains 9 stimuli listing the names of universities that made up the Big Nine conference.
}
\usage{data(big.nine)}
\format{
  A data frame with 9 observations on the following variable.
  \describe{
    \item{\code{University}}{is a data.frame with the following stimuli: 
	
	University of Chicago
	
	University of Illinois
	
	University of Michigan
	
	University of Minnesota
	
	Northwestern University
	
	Purdue University
	
	University of Wisconsin
	
	Indiana University
	
	University of Iowa }
  }
}
\examples{
data(big.nine)
ross.ordering(big.nine)}

