\name{Cloud}
\alias{Cloud}
\docType{data}
\title{
Cloud
}
\description{
A data frame with 1024 observations on the following 10 variables.
}
\usage{data("Cloud")}
\format{
A data frame with 1024 observations on the following 10 variables.
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{x4}}{a numeric vector}
    \item{\code{x5}}{a numeric vector}
    \item{\code{x6}}{a numeric vector}
    \item{\code{x7}}{a numeric vector}
    \item{\code{x8}}{a numeric vector}
    \item{\code{x9}}{a numeric vector}
    \item{\code{x10}}{a numeric vector}
  }
}
\details{
The data sets we propose to analyse are constituted of 1024 vectors, each vector includes 10 parameters. You can think of it as a 1024*10 matrix.
}
\source{
The Cloud data set comes from the UCI database.
}
\references{
NA
}
\examples{
data(Cloud)
## maybe str(Cloud) ; plot(Cloud) ...
}
\keyword{datasets}
