\name{HTRU}
\alias{HTRU}
\docType{data}
\title{
HTRU
}
\description{
A data frame with 10000 observations on the following 9 variables.
}
\usage{data("HTRU")}
\format{
  A data frame with 10000 observations on the following 9 variables.
  \describe{
    \item{\code{A}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{C}}{a numeric vector}
    \item{\code{D}}{a numeric vector}
    \item{\code{E}}{a numeric vector}
    \item{\code{F}}{a numeric vector}
    \item{\code{G}}{a numeric vector}
    \item{\code{H}}{a numeric vector}
    \item{\code{I}}{a numeric vector}
  }
}
\details{
Pulsar candidates collected during the HTRU survey. Pulsars are a type of star, of considerable scientific interest. Candidates must be classified in to pulsar and non-pulsar classes to aid discovery.
}
\source{
The HTRU data set comes from the UCI database.
}
\references{
NA
}
\examples{
data(HTRU)
## maybe str(HTRU) ; plot(HTRU) ...
}
\keyword{datasets}
