\name{politicians}
\alias{politicians}
\docType{data}
\title{Example dataset for Ross Ordering Pairing}
\description{
Dataset contains 8 stimuli listing 8 names of politicians.
}
\usage{data(politicians)}
\format{
  A data frame with 8 observations on the following variable.
  \describe{
    \item{\code{Politicians}}{is a data.frame with the following stimuli: 
	
	Anderson
	
	Reagan
	
	Carter 
	
	Mondale 
	
	Bush
	
	Lucey 
	
	Oneil 
	
	Baker }
  }
}
\examples{
data(politicians)
ross.ordering(politicians)}

