\name{Wine}
\alias{Wine}
\docType{data}
\title{
Wine
}
\description{
A data frame with 177 observations on the following 13 variables.
}
\usage{data("Wine")}
\format{
A data frame with 177 observations on the following 13 variables.
  \describe{
    \item{\code{X14.23}}{a numeric vector}
    \item{\code{X1.71}}{a numeric vector}
    \item{\code{X2.43}}{a numeric vector}
    \item{\code{X15.6}}{a numeric vector}
    \item{\code{X127}}{a numeric vector}
    \item{\code{X2.8}}{a numeric vector}
    \item{\code{X3.06}}{a numeric vector}
    \item{\code{X.28}}{a numeric vector}
    \item{\code{X2.29}}{a numeric vector}
    \item{\code{X5.64}}{a numeric vector}
    \item{\code{X1.04}}{a numeric vector}
    \item{\code{X3.92}}{a numeric vector}
    \item{\code{X1065}}{a numeric vector}
  }
}
\details{
These data are the results of a chemical analysis of wines grown in the same region in Italy but derived from three different cultivars. The analysis determined the quantities of 13 constituents found in each of the three types of wines. 
}
\source{
The Wine data set comes from the UCI database.
}
\references{
NA
}
\examples{
data(Wine)
## maybe str(Wine) ; plot(Wine) ...
}
\keyword{datasets}